Move the cost report export stack into the CostReports namespace
