+++ modules/reporting/app/workers/cost_reports/pdf/schema.json
